infrastructure/nvidia-parakeet-model-fine-tuning/requirements.txt
# Stage 1: Core packages

numpy>=1.22
setuptools>=70.0.0

# Stage 2: Basic dependencies
protobuf~=4.24.4
python-dateutil
ruamel.yaml
scikit-learn
tensorboard
text-unidecode
tqdm>=4.41.0
wget
wrapt
cloudpickle
numba
pyrubberband

# Stage 3: ML/AI packages

transformers==4.52.0
accelerate==1.7.0
torchmetrics>=0.11.0
onnx>=1.7.0
huggingface_hub>=0.24
datasets==3.6.0
# Stage 4: Configuration packages
hydra-core>1.3,<=1.3.2
omegaconf<=2.3
wandb
mlflow
peft

# Stage 5: Data processing
webdataset>=0.2.86
fsspec==2024.12.0

# Stage 6: CUDA packages
cuda-python>=12.3


# Stage 7: Complex packages
nemo_toolkit[all]
deepspeed